[refactor] adds typescript to Auth components
